# doc/api/openapi.yaml
"openapi": "3.1.0"
"info":
  "title": "PixelPast"
  "version": "0.1.0"
"paths":
...
  "/api/social/graph":
    "get":
      "tags":
        - "social"
      "summary": "Get social graph projection"
      "description": "Return a person co-occurrence graph derived from qualifying assets within the resolved inclusive UTC date range."
      "operationId": "get_social_graph_api_social_graph_get"
      "parameters":
        -
          "name": "start"
          "in": "query"
          "required": false
          "schema":
            "anyOf":
              -
                "type": "string"
                "format": "date"
              -
                "type": "null"
            "description": "Inclusive UTC start date for the requested social-graph window. Must be provided together with end."
            "title": "Start"
          "description": "Inclusive UTC start date for the requested social-graph window. Must be provided together with end."
          "examples":
            "trip_window_start":
              "summary": "Graph window start"
              "value": "2026-07-01"
        -
          "name": "end"
          "in": "query"
          "required": false
          "schema":
            "anyOf":
              -
                "type": "string"
                "format": "date"
              -
                "type": "null"
            "description": "Inclusive UTC end date for the requested social-graph window. Must be provided together with start."
            "title": "End"
          "description": "Inclusive UTC end date for the requested social-graph window. Must be provided together with start."
          "examples":
            "trip_window_end":
              "summary": "Graph window end"
              "value": "2026-07-31"
        -
          "name": "person_ids"
          "in": "query"
          "required": false
          "schema":
            "type": "array"
            "items":
              "type": "integer"
            "description": "Repeatable seed person identifiers. Only assets containing at least one selected person qualify when this filter is present."
            "default": []
            "title": "Person Ids"
          "description": "Repeatable seed person identifiers. Only assets containing at least one selected person qualify when this filter is present."
          "examples":
            "anna":
              "summary": "Seed the graph with Anna Becker"
              "value": 7
        -
          "name": "max_people_per_asset"
          "in": "query"
          "required": false
          "schema":
            "type": "integer"
            "maximum": 30
            "minimum": 2
            "description": "Ignore assets that contain more than this many linked people to avoid oversized group shots dominating the graph."
            "default": 10
            "title": "Max People Per Asset"
          "description": "Ignore assets that contain more than this many linked people to avoid oversized group shots dominating the graph."
          "examples":
            "medium_group_cap":
              "summary": "Ignore large group shots"
              "value": 8
        -
          "name": "tag_paths"
          "in": "query"
          "required": false
          "schema":
            "type": "array"
            "items":
              "type": "string"
            "description": "Reserved filter parameter. The current social graph contract rejects tag-based filtering."
            "default": []
            "title": "Tag Paths"
          "description": "Reserved filter parameter. The current social graph contract rejects tag-based filtering."
          "examples":
            "unsupported_tag":
              "summary": "Rejected tag filter example"
              "value": "travel/italy/venice"
        -
          "name": "location_geometry"
          "in": "query"
          "required": false
          "schema":
            "anyOf":
              -
                "type": "string"
              -
                "type": "null"
            "description": "Reserved filter parameter. The current social graph contract rejects geometry-based filtering."
            "title": "Location Geometry"
          "description": "Reserved filter parameter. The current social graph contract rejects geometry-based filtering."
          "examples":
            "unsupported_geometry":
              "summary": "Rejected geometry filter example"
              "value": "bbox:12.285,45.423,12.380,45.465"
        -
          "name": "distance_latitude"
          "in": "query"
          "required": false
          "schema":
            "anyOf":
              -
                "type": "number"
                "maximum": 90
                "minimum": -90
              -
                "type": "null"
            "description": "Reserved filter parameter. The current social graph contract rejects radial distance filtering."
            "title": "Distance Latitude"
          "description": "Reserved filter parameter. The current social graph contract rejects radial distance filtering."
          "examples":
            "unsupported_distance_lat":
              "summary": "Rejected distance filter latitude"
              "value": 45.4371
        -
          "name": "distance_longitude"
          "in": "query"
          "required": false
          "schema":
            "anyOf":
              -
                "type": "number"
                "maximum": 180
                "minimum": -180
              -
                "type": "null"
            "description": "Reserved filter parameter. The current social graph contract rejects radial distance filtering."
            "title": "Distance Longitude"
          "description": "Reserved filter parameter. The current social graph contract rejects radial distance filtering."
          "examples":
            "unsupported_distance_lon":
              "summary": "Rejected distance filter longitude"
              "value": 12.3327
        -
          "name": "distance_radius_meters"
          "in": "query"
          "required": false
          "schema":
            "anyOf":
              -
                "type": "integer"
                "minimum": 1
              -
                "type": "null"
            "description": "Reserved filter parameter. The current social graph contract rejects radial distance filtering."
            "title": "Distance Radius Meters"
          "description": "Reserved filter parameter. The current social graph contract rejects radial distance filtering."
          "examples":
            "unsupported_distance_radius":
              "summary": "Rejected distance filter radius"
              "value": 5000
        -
          "name": "filename_query"
          "in": "query"
          "required": false
          "schema":
            "anyOf":
              -
                "type": "string"
                "minLength": 1
              -
                "type": "null"
            "description": "Reserved filter parameter. The current social graph contract rejects filename-based filtering."
            "title": "Filename Query"
          "description": "Reserved filter parameter. The current social graph contract rejects filename-based filtering."
          "examples":
            "unsupported_filename":
              "summary": "Rejected filename filter example"
              "value": "IMG_20260703"
      "responses":
        "200":
          "description": "Person nodes and weighted links for the filtered social graph."
          "content":
            "application/json":
              "schema":
                "$ref": "#/components/schemas/SocialGraphResponse"
              "examples":
                "trip_companions":
                  "summary": "Social graph for a filtered travel window"
                  "value":
                    "persons":
                      -
                        "id": 7
                        "name": "Anna Becker"
                        "occurrence_count": 12
                      -
                        "id": 12
                        "name": "Milo Tan"
                        "occurrence_count": 8
                      -
                        "id": 19
                        "name": "Nora Rossi"
                        "occurrence_count": 5
                    "links":
                      -
                        "person_ids":
                          - 7
                          - 12
                        "weight": 6
                        "affinity": 0.75
                      -
                        "person_ids":
                          - 7
                          - 19
                        "weight": 3
                        "affinity": 0.428571
        "400":
          "description": "The request was syntactically valid but violated a documented PixelPast API contract constraint."
          "content":
            "application/json":
              "examples":
                "invalid_range":
                  "summary": "Start date after end date"
                  "value":
                    "detail": "start must be less than or equal to end"
                "partial_range":
                  "summary": "Only one range boundary provided"
                  "value":
                    "detail": "start and end must both be provided together"
                "unsupported_filters":
                  "summary": "Rejected tag and filename filters"
                  "value":
                    "detail": "unsupported social graph filters: filename_query, tag_paths; supported filters: start, end, person_ids, max_people_per_asset"
              "schema":
                "$ref": "#/components/schemas/ApiErrorResponse"
        "422":
          "description": "The request could not be parsed or failed declared parameter validation."
          "content":
            "application/json":
              "examples":
                "invalid_date":
                  "summary": "Unparseable ISO date"
                  "value":
                    "detail":
                      -
                        "type": "date_from_datetime_parsing"
                        "loc":
                          - "query"
                          - "start"
                        "msg": "Input should be a valid date or datetime, input is too short"
                        "input": "2026-02"
                "invalid_path_date":
                  "summary": "Invalid path day format"
                  "value":
                    "detail":
                      -
                        "type": "date_from_datetime_parsing"
                        "loc":
                          - "path"
                          - "day"
                        "msg": "Input should be a valid date or datetime, invalid date separator"
                        "input": "2026/07/14"
              "schema":
                "$ref": "#/components/schemas/ApiValidationErrorResponse"
"components":
  "schemas":
    "ApiErrorResponse":
      "properties":
        "detail":
          "type": "string"
          "title": "Detail"
      "type": "object"
      "required":
        - "detail"
      "title": "ApiErrorResponse"
      "description": "Error payload for contract-level request rejections."
    "ApiValidationErrorItem":
      "properties":
        "type":
          "type": "string"
          "title": "Type"
        "loc":
          "items":
            "anyOf":
              -
                "type": "string"
              -
                "type": "integer"
          "type": "array"
          "title": "Loc"
        "msg":
          "type": "string"
          "title": "Msg"
        "input":
          "anyOf":
            - {}
            -
              "type": "null"
          "title": "Input"
      "type": "object"
      "required":
        - "type"
        - "loc"
        - "msg"
      "title": "ApiValidationErrorItem"
      "description": "One FastAPI request validation error entry."
    "ApiValidationErrorResponse":
      "properties":
        "detail":
          "items":
            "$ref": "#/components/schemas/ApiValidationErrorItem"
          "type": "array"
          "title": "Detail"
      "type": "object"
      "required":
        - "detail"
      "title": "ApiValidationErrorResponse"
      "description": "Error payload returned when declared request validation fails."
...
    "SocialGraphLink":
      "properties":
        "person_ids":
          "items":
            "type": "integer"
          "type": "array"
          "title": "Person Ids"
        "weight":
          "type": "integer"
          "title": "Weight"
        "affinity":
          "type": "number"
          "title": "Affinity"
      "type": "object"
      "required":
        - "person_ids"
        - "weight"
        - "affinity"
      "title": "SocialGraphLink"
      "description": "Weighted unordered person-pair edge in the social graph."
    "SocialGraphPerson":
      "properties":
        "id":
          "type": "integer"
          "title": "Id"
        "name":
          "type": "string"
          "title": "Name"
        "occurrence_count":
          "type": "integer"
          "title": "Occurrence Count"
      "type": "object"
      "required":
        - "id"
        - "name"
        - "occurrence_count"
      "title": "SocialGraphPerson"
      "description": "Person node exposed by the social-graph API contract."
    "SocialGraphResponse":
      "properties":
        "persons":
          "items":
            "$ref": "#/components/schemas/SocialGraphPerson"
          "type": "array"
          "title": "Persons"
        "links":
          "items":
            "$ref": "#/components/schemas/SocialGraphLink"
          "type": "array"
          "title": "Links"
      "type": "object"
      "required":
        - "persons"
        - "links"
      "title": "SocialGraphResponse"
      "description": "Canonical social-graph projection response."
...